requirements.txt
google-generativeai
yfinance
pytz
playwright
beautifulsoup4
